% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECBSDW.R
\docType{package}
\name{ECBSDW}
\alias{ECBSDW}
\alias{ECBSDW-package}
\title{A package for downloading time-series from the ECB Statistical Data Warehouse.}
\description{
The workhorse function of the package is \code{getECBSDW()}. This function returns a
dataframe containing the requested time-series. The code behind this package also serves as a backend
to an EViews add-in created by the author.
}
\section{ECBSDW functions}{

\code{getECBSDW()}
}

