% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECBSDW.R
\name{getECBSDW}
\alias{getECBSDW}
\title{Get a time-series from the ECB SDW}
\usage{
getECBSDW(series_key)
}
\arguments{
\item{series_key}{The series key for the time-series to be downloaded}
}
\description{
Downloads a time-series from ECB SDW into R
}
\details{
This function uses a very simple interface with the ECB SDW and does not use the ECB SDW API, unlike some other R packages.
}
\author{
Graeme Walsh
}
